projects:ad9172: Use iio_app
Refactor project to use common iio_app
