+++ projects/ad9172/src/parameters.h
@@ -55,6 +55,7 @@
 
 #define UART_DEVICE_ID				XPAR_XUARTPS_0_DEVICE_ID
 #define UART_IRQ_ID				XPAR_XUARTPS_1_INTR
+#define UART_BAUDRATE                           921600
 
 #define INTC_DEVICE_ID				XPAR_SCUGIC_SINGLE_DEVICE_ID
